refactor: Parse를 하는기능 모두 HttpParser로 변경
